notebooks/HelloWorldFunc/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
certifi==2020.6.20
cycler==0.10.0
joblib==0.17.0
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.3
seaborn==0.11.0
six==1.15.0
threadpoolctl==2.1.0
